+++ eventlog-live/src/GHC/Eventlog/Live/Machine/Core.hs
@@ -9,35 +9,27 @@ Portability : portable
 module GHC.Eventlog.Live.Machine.Core (
   -- * Ticks
   Tick (..),
+  batchByTickList,
+  batchByTicksList,
   batchByTick,
   batchByTicks,
   batchToTick,
   batchToTicks,
-  batchByTickList,
-  batchByTicksList,
-  batchListToTick,
-  batchListToTicks,
   dropTick,
   onlyTick,
-  aggregateByTick,
-  aggregateByTicks,
   liftTick,
-  liftBatch,
-
-  -- * Debug
-  counterBy,
-  counterByTick,
 
   -- * Routers
   liftRouter,
 
   -- * Event sorting
   sortByBatch,
-  sortByBatchTick,
+  sortByTick,
 
   -- * Delimiting
-  between,
   delimit,
+  betweenEach,
+  betweenFirst,
 
   -- * Validation
   validateInput,
@@ -47,7 +39,7 @@ module GHC.Eventlog.Live.Machine.Core (
 import Control.Monad (replicateM_, when)
 import Control.Monad.IO.Class (MonadIO (..))
 import Data.DList qualified as D
-import Data.Foldable (for_)
+import Data.Foldable (Foldable (..), for_)
 import Data.Function (on)
 import Data.Functor ((<&>))
 import Data.HashMap.Strict (HashMap)
@@ -57,13 +49,16 @@ import Data.List qualified as L
 import Data.Machine (Is (..), MachineT (..), Moore (..), PlanT, Process, ProcessT, Step (..), asParts, await, construct, encased, mapping, repeatedly, starve, stopped, yield, (~>))
 import Data.Maybe (fromMaybe)
 import Data.Semigroup (Max (..))
-import Data.Text (Text)
 import Data.Text qualified as T
+import Data.Void (Void)
 import GHC.Eventlog.Live.Logger (logDebug, logError, logWarning)
-import GHC.Eventlog.Live.Verbosity (Verbosity, verbosityDebug, verbosityError, verbosityWarning)
-import GHC.RTS.Events (Event (..), Timestamp)
-import GHC.RTS.Events qualified as E
+import GHC.Eventlog.Live.Verbosity (Verbosity, verbosityError, verbosityWarning)
 import Text.Printf (printf)
+
+{- $setup
+>>> import Data.Machine ((<~), auto, echo, fold, taking, run, runT_, scan, source)
+>>> import Data.Semigroup (Sum (..))
+-}
 
 -------------------------------------------------------------------------------
 -- Ticks
@@ -79,16 +74,43 @@ data Tick a = Item !a | Tick
   deriving (Eq, Functor, Foldable, Traversable, Show)
 
 {- |
-This machine batches all items between two ticks into a list.
+Batches items to lists.
+
+The process @`batchByTickList`@ consumes a stream of items and ticks.
+It preserves ticks but batches items between ticks to lists.
+
+__Warning:__ This process does not yield empty batches.
+
+==== __Examples__
+
+>>> run $ batchByTickList <~ source [Item 1,Item 2,Tick,Item 3,Tick,Item 4,Item 5,Tick,Item 6,Tick]
+[[1,2],[3],[4,5],[6]]
+
+>>> run $ batchByTickList <~ source [Item 1,Item 2,Tick,Tick]
+[[1,2]]
 -}
 batchByTickList :: Process (Tick a) [a]
 batchByTickList =
   mapping (fmap D.singleton)
     ~> batchByTick
+    ~> dropTick
     ~> mapping D.toList
 
 {- |
-Generalised version of `batchByTickList`.
+Batches items for a given number of ticks to lists.
+
+The process @`batchByTicksList` n@ consumes a stream of items and ticks.
+It preserves ticks but batches items for @n@ ticks to lists and yields the batch before the @n@'th tick.
+
+__Warning:__ This process does not yield empty batches.
+
+==== __Examples__
+
+>>> run $ batchByTicksList 2 <~ source [Item 1,Item 2,Tick,Item 3,Tick,Item 4,Item 5,Tick,Item 6,Tick]
+[[1,2,3],[4,5,6]]
+
+>>> run $ batchByTicksList 2 <~ source [Item 1,Item 2,Tick,Tick]
+[[1,2]]
 -}
 batchByTicksList ::
   -- | The number of ticks per batch.
@@ -97,64 +119,104 @@ batchByTicksList ::
 batchByTicksList ticks =
   mapping (fmap D.singleton)
     ~> batchByTicks ticks
+    ~> dropTick
     ~> mapping D.toList
 
 {- |
-Generalised version of `batchByTickList`.
+Batches items via their `Semigroup` instance.
+
+The process @`batchByTick`@ consumes a stream of items and ticks.
+It preserves ticks but batches items between ticks using `sconcat`.
+
+==== __Examples__
+
+>>> run $ batchByTick <~ source [Item [1],Item [2],Tick,Item [3],Tick,Item [4],Item [5],Tick,Item [6],Tick]
+[Item [1,2],Tick,Item [3],Tick,Item [4,5],Tick,Item [6],Tick]
+
+>>> run $ batchByTick <~ source [Item (Sum 1),Item (Sum 2),Tick,Item (Sum 3),Tick,Item (Sum 4),Item (Sum 5),Tick,Item (Sum 6),Tick]
+[Item (Sum {getSum = 3}),Tick,Item (Sum {getSum = 3}),Tick,Item (Sum {getSum = 9}),Tick,Item (Sum {getSum = 6}),Tick]
 -}
 batchByTick ::
   forall a.
-  (Monoid a) => Process (Tick a) a
+  (Monoid a) => Process (Tick a) (Tick a)
 batchByTick = batchByTicks 1
 
 {- |
-Generalised version of `batchByTick`.
+Batches items for a given number of ticks via their `Semigroup` instance.
+
+The process @`batchByTicks` n@ consumes a stream of items and ticks.
+It preserves ticks but batches items for @n@ ticks using `sconcat` and yields the batch before the @n@'th tick.
+
+==== __Examples__
+
+>>> run $ batchByTicks 2 <~ source [Item [1],Item [2],Tick,Item [3],Tick,Item [4],Item [5],Tick,Item [6],Tick]
+[Tick,Item [1,2,3],Tick,Tick,Item [4,5,6],Tick]
+
+>>> run $ batchByTicks 2 <~ source [Item (Sum 1),Item (Sum 2),Tick,Item (Sum 3),Tick,Item (Sum 4),Item (Sum 5),Tick,Item (Sum 6),Tick]
+[Tick,Item (Sum {getSum = 6}),Tick,Tick,Item (Sum {getSum = 15}),Tick]
 -}
 batchByTicks ::
   forall a.
-  (Monoid a) =>
+  (Semigroup a) =>
   -- | The number of ticks per batch.
   Int ->
-  Process (Tick a) a
+  Process (Tick a) (Tick a)
 batchByTicks ticks = batchByTicksWith ticks mempty
  where
   batchByTicksWith ::
     forall m.
     (Monad m) =>
     Int ->
-    a ->
-    MachineT m (Is (Tick a)) a
-  batchByTicksWith ticksRemaining acc = MachineT $ pure $ Await onNext Refl onStop
+    Maybe a ->
+    MachineT m (Is (Tick a)) (Tick a)
+  batchByTicksWith ticksRemaining maybeAcc =
+    MachineT $ pure $ Await onNext Refl onStop
    where
-    onNext :: Tick a -> MachineT m (Is (Tick a)) a
+    yieldItem :: a -> ProcessT m (Tick a) (Tick a) -> ProcessT m (Tick a) (Tick a)
+    yieldItem a = MachineT . pure . Yield (Item a)
+
+    yieldTick :: ProcessT m (Tick a) (Tick a) -> ProcessT m (Tick a) (Tick a)
+    yieldTick = MachineT . pure . Yield Tick
+
+    onNext :: Tick a -> MachineT m (Is (Tick a)) (Tick a)
     onNext = \case
-      Item a -> batchByTicksWith ticksRemaining (a <> acc)
+      Item a -> batchByTicksWith ticksRemaining (maybeAcc <> Just a)
       Tick
-        | ticksRemaining <= 1 -> MachineT $ pure $ Yield acc $ batchByTicksWith ticks mempty
-        | otherwise -> batchByTicksWith (ticksRemaining - 1) acc
-    onStop :: MachineT m (Is (Tick a)) a
-    onStop = MachineT $ pure $ Yield acc stopped
+        | ticksRemaining <= 1 ->
+            -- Yield an `Item` if any items were accumulated.
+            maybe id yieldItem maybeAcc $
+              -- Yield the `Tick`.
+              yieldTick $
+                -- Continue with the initial state.
+                batchByTicksWith ticks Nothing
+        | otherwise ->
+            -- Yield the `Tick`.
+            yieldTick $
+              -- Continue with one fewer tick remaining.
+              batchByTicksWith (ticksRemaining - 1) maybeAcc
+
+    onStop :: MachineT m (Is (Tick a)) (Tick a)
+    onStop =
+      -- Yield an `Item` if any items were accumulated.
+      maybe id yieldItem maybeAcc $
+        -- Stop.
+        stopped
 
 {- |
-This machine streams a list of items into a series of items
-separated by ticks.
--}
-batchListToTick ::
-  Process [a] (Tick a)
-batchListToTick = batchToTick
+Streams batches of items as items and ticks.
 
-{- |
-This machine streams a list of items into a series of items
-separated by the given number of ticks.
--}
-batchListToTicks ::
-  -- | The number of ticks per batch.
-  Int ->
-  Process [a] (Tick a)
-batchListToTicks = batchToTicks
+The process @`batchToTick`@ consumes a stream of batches of items
+and produces a stream of items and with a tick separating each batch.
 
-{- |
-Generalised version of `batchListToTick`.
+__Warning:__ The process @`batchByTickList` `~>` `batchToTick`@ is not the identity!
+
+==== __Examples__
+
+>>> run $ batchToTick <~ source [[1,2],[3,4]]
+[Item 1,Item 2,Tick,Item 3,Item 4,Tick]
+
+>>> run $ batchToTick <~ batchByTickList <~ source [Item 1,Item 2,Tick,Tick]
+[Item 1,Item 2,Tick]
 -}
 batchToTick ::
   (Foldable f) =>
@@ -162,9 +224,23 @@ batchToTick ::
 batchToTick = batchToTicks 1
 
 {- |
-Generalised version of `batchToTicks`.
+Streams batches of items as items and ticks.
 
-__Warning:__ This machine yields all items between the penultimate and the last tick.
+The process @`batchToTicks` n@ consumes a stream of lists of items
+and produces a stream of items and with @n@ ticks separating each list.
+
+To mirror the batches created by @`batchByTicks` n@, each list is
+opened with @n - 1@ ticks and closed by the @n@'th tick.
+
+__Warning:__ The process @`batchByTicks` n `~>` `batchToTicks` n@ is not the identity!
+
+==== __Examples__
+
+>>> run $ batchToTicks 2 <~ source [[1,2],[3,4]]
+[Tick,Item 1,Item 2,Tick,Tick,Item 3,Item 4,Tick]
+
+>>> run $ batchToTicks 2 <~ batchByTicksList 2 <~ source [Item 1,Item 2,Tick,Tick]
+[Tick,Item 1,Item 2,Tick]
 -}
 batchToTicks ::
   (Foldable f) =>
@@ -195,76 +271,6 @@ onlyTick =
       Tick -> yield ()
       Item{} -> pure ()
 
-{- |
-This machine aggregates a value by tick.
-
-The difference between `batchByTick` and `aggregateByTick` is that
-`batchByTick` yields a batch on every tick whereas
-`aggregateByTick` only yields a batch if there were any values.
--}
-aggregateByTick :: (Semigroup a) => Process (Tick a) a
-aggregateByTick =
-  mapping (fmap Just)
-    ~> batchByTick
-    ~> asParts
-
-{- |
-Generalised variant of `aggregateByTick`.
--}
-aggregateByTicks ::
-  (Semigroup a) =>
-  -- | The number of ticks per batch.
-  Int ->
-  Process (Tick a) a
-aggregateByTicks ticks =
-  mapping (fmap Just)
-    ~> batchByTicks ticks
-    ~> asParts
-
--------------------------------------------------------------------------------
--- Machine combinators
--------------------------------------------------------------------------------
-
-{- |
-This machine counts the number of inputs it received,
-using the given function, and logs this value.
--}
-counterBy ::
-  forall m a x.
-  (MonadIO m) =>
-  Verbosity ->
-  Text ->
-  (a -> Word) ->
-  ProcessT m a x
-counterBy verbosity label counter
-  | verbosityDebug >= verbosity = repeatedly go
-  | otherwise = stopped
- where
-  go :: PlanT (Is a) x m ()
-  go =
-    await >>= \a ->
-      logDebug verbosity (T.pack (show (counter a)) <> " " <> label)
-
-{- |
-This machine counts the number of inputs it received,
-and logs this value on every tick.
--}
-counterByTick ::
-  forall m a x.
-  (MonadIO m) =>
-  Verbosity ->
-  Text ->
-  ProcessT m (Tick a) x
-counterByTick verbosity label
-  | verbosityDebug >= verbosity = construct $ go 0
-  | otherwise = stopped
- where
-  go :: Word -> PlanT (Is (Tick a)) x m ()
-  go count =
-    await >>= \case
-      Item _ -> go (count + 1)
-      Tick -> logDebug verbosity (T.pack (show count) <> " " <> label) >> go 0
-
 -------------------------------------------------------------------------------
 -- Machine combinators
 -------------------------------------------------------------------------------
@@ -282,6 +288,14 @@ Constructs the following machine:
   [ Tick a ]                              [ Tick b ]
            └─(if Item)─( ProcessT m a b )─┘
 @
+
+==== __Examples__
+
+>>> run $ liftTick (mapping (+1)) <~ source [Item 1,Tick,Item 2,Item 3,Tick,Tick]
+[Item 2,Tick,Item 3,Item 4,Tick,Tick]
+
+>>> run $ liftTick (scan (+) 0) <~ source [Item 1,Tick,Item 2,Item 3,Tick,Tick]
+[Item 0,Item 1,Tick,Item 3,Item 6,Tick,Tick]
 -}
 liftTick ::
   (Monad m) =>
@@ -310,126 +324,123 @@ liftTick m =
           onStop' = liftTick onStop
 
 --------------------------------------------------------------------------------
--- Lift a machine to a machine that operates on batches
-
-{- |
-Lift a machine that processes @a@s into @b@s to a machine that processes
-batches of @a@s into batches of @b@s.
--}
-liftBatch ::
-  forall m a b.
-  (Monad m) =>
-  ProcessT m a b ->
-  ProcessT m [a] [b]
-liftBatch = MachineT . running [] []
- where
-  -- The parent machine is running the child machine with the current batch.
-  running :: [a] -> [b] -> ProcessT m a b -> m (Step (Is [a]) [b] (ProcessT m [a] [b]))
-  running as bs m =
-    runMachineT m >>= \case
-      Stop ->
-        pure Stop
-      Yield b k ->
-        running as (b : bs) k
-      Await (onNext :: t -> ProcessT m a b) Refl onStop ->
-        pure $ Yield (reverse bs) $ MachineT $ awaiting as onNext onStop
-
-  -- The parent machine is awaiting new input.
-  awaiting :: [a] -> (a -> ProcessT m a b) -> ProcessT m a b -> m (Step (Is [a]) [b] (ProcessT m [a] [b]))
-  awaiting (a : as) onNext _onStop = running as [] $ onNext a
-  awaiting [] onNext onStop = pure $ Await onNext' Refl onStop'
-   where
-    onNext' :: [a] -> ProcessT m [a] [b]
-    onNext' as = MachineT $ awaiting as onNext onStop
-    onStop' :: ProcessT m [a] [b]
-    onStop' = exhausting onStop
-
-  -- The parent machine is exhausting the child machine to gather its output.
-  exhausting :: ProcessT m a b -> ProcessT m x [b]
-  exhausting = MachineT . go []
-   where
-    go :: [b] -> ProcessT m a b -> m (Step (Is x) [b] (ProcessT m x [b]))
-    go bs m =
-      runMachineT m >>= \case
-        Stop ->
-          pure Stop
-        Yield b k ->
-          go (b : bs) k
-        Await _onNext _Refl onStop ->
-          pure $ Yield (reverse bs) $ MachineT $ go [] onStop
-
---------------------------------------------------------------------------------
 -- Construct a processor that spawns a separate child processor for each measure
 
 {- |
-Spawn a copy of a machine for each "measure".
+Spawns a process for each measure.
 
 Constructs the following machine:
 
 @
-    ┌─────(if measure == k0)─( ProcessT m a b )────┐
-  [ a ] ──(if measure == ..)─( ProcessT m a b )─ [ b ]
-    └─────(if measure == kN)─( ProcessT m a b )────┘
+    ┌─────(if measure == k0)─( spawn k0 :: ProcessT m a b )────┐
+  [ a ] ──(if measure == ..)─( spawn .. :: ProcessT m a b )─ [ b ]
+    └─────(if measure == kN)─( spawn kN :: ProcessT m a b )────┘
 @
 
-__Warning:__ The router does not currently garbage-collect terminated child processors.
+If the spawned process for some measure stops,
+then all future inputs for that measure are ignored.
+
+__Warning:__
+The router process holds on to a reference to each measure @i@ for each child
+process, even after that child process has stopped.
+
+==== __Examples__
+
+>>> run $ liftRouter (Just . even) (\case {True -> mapping (+1); False -> echo}) <~ source [1,2,3,4,5]
+[1,3,3,5,5]
+
+>>> run $ liftRouter (Just . even) (\case {True -> echo; False -> stopped}) <~ source [1,2,3,4,5]
+[2,4]
+
+>>> run $ liftRouter (Just . even) (\case {True -> echo; False -> taking 1}) <~ source [1,2,3,4,5]
+[1,2,4]
 -}
 liftRouter ::
-  forall m k a b.
-  (MonadIO m, Hashable k) =>
+  forall m i a b.
+  (Monad m, Hashable i) =>
   -- | Function to measure.
-  (a -> Maybe k) ->
+  (a -> Maybe i) ->
   -- | Function to spawn child processors.
-  (k -> ProcessT m a b) ->
+  (i -> ProcessT m a b) ->
   ProcessT m a b
 liftRouter measure spawn = awaiting M.empty
  where
-  awaiting :: HashMap k (ProcessT m a b) -> ProcessT m a b
+  awaiting :: HashMap i (Child (ProcessT m a b)) -> ProcessT m a b
   awaiting st = MachineT . pure $ Await onNext Refl onStop
    where
     onNext :: a -> MachineT m (Is a) b
     onNext a = case measure a of
       Nothing -> awaiting st
-      Just k -> provideThen a m $ \m' -> awaiting (M.insert k m' st)
-       where
-        m = fromMaybe (spawn k) (M.lookup k st)
-    onStop :: MachineT m (Is a) b
-    onStop = foldr starve stopped (M.elems st)
+      Just i ->
+        case fromMaybe (ChildRunning $ spawn i) (M.lookup i st) of
+          ChildRunning p ->
+            provideThen a p $ \p' ->
+              let !st' = M.insert i p' st
+               in awaiting st'
+          ChildStopped ->
+            awaiting st
 
-  provideThen :: a -> ProcessT m a b -> (ProcessT m a b -> ProcessT m a b) -> ProcessT m a b
-  provideThen a m k =
+    onStop :: MachineT m (Is a) b
+    onStop = foldr starve stopped (concatMap toList . M.elems $ st)
+
+  provideThen :: a -> ProcessT m a b -> (Child (ProcessT m a b) -> ProcessT m a b) -> ProcessT m a b
+  provideThen a p k =
     MachineT $
-      runMachineT m >>= \case
-        Stop -> runMachineT (k stopped)
-        Yield o m' -> pure (Yield o (provideThen a m' k))
+      runMachineT p >>= \case
+        Stop -> runMachineT (k ChildStopped)
+        Yield o p' -> pure (Yield o (provideThen a p' k))
         Await onNext Refl _onStop -> runMachineT (exhaustThen (onNext a) k)
 
-  exhaustThen :: ProcessT m a b -> (ProcessT m a b -> ProcessT m a b) -> ProcessT m a b
-  exhaustThen m k =
+  exhaustThen :: ProcessT m a b -> (Child (ProcessT m a b) -> ProcessT m a b) -> ProcessT m a b
+  exhaustThen p k =
     MachineT $
-      runMachineT m >>= \case
-        Yield o m' -> pure (Yield o (k m'))
-        m' -> runMachineT (k (encased m'))
+      runMachineT p >>= \case
+        Yield o p' -> pure (Yield o (k $ ChildRunning p'))
+        p' -> runMachineT (k (ChildRunning $ encased p'))
+
+{- |
+Internal helper.
+A wrapper for child processes spawned by `liftRouter`.
+-}
+data Child a
+  = ChildRunning !a
+  | ChildStopped
+  deriving (Functor, Foldable)
 
 -------------------------------------------------------------------------------
 -- Event stream sorting
 -------------------------------------------------------------------------------
 
 {- |
-Reorder events respecting ticks.
+Sort items in successive batches.
 
-This machine caches two batches worth of events, sorts them together,
-and then yields only those events whose timestamp is less than or equal
-to the maximum of the first batch.
+If the maximum key in batch @i@ is guaranteed to be smaller than the minimum
+key in batch @i + 2@, this process produces a totally ordered stream of items.
+
+The process @`sortByBatch` key@ caches two batches of items, sorts them
+together, and yields only those items whose key is less than or equal
+to the maximum key in the first batch.
+
+==== __Examples__
+
+>>> run $ sortByBatch @Int id <~ source [[1,4],[7,2,3,5],[6,8]]
+[[1,2,3,4],[5,6,7],[8]]
+
+>>> run $ sortByBatch @Int id <~ source [[1,7],[4,2,3,5],[6,8]]
+[[1,2,3,4,5,7],[6,8]]
 -}
 sortByBatch ::
-  forall m a.
-  (Monad m) =>
-  (a -> Timestamp) ->
-  ProcessT m [a] [a]
-sortByBatch timestamp = sortByBatchWith Nothing
+  forall a k.
+  (Bounded k, Ord k) =>
+  (a -> k) ->
+  Process [a] [a]
+sortByBatch key = sortByBatchWith Nothing
  where
-  sortByBatchWith :: Maybe [a] -> ProcessT m [a] [a]
+  sortByBatchWith ::
+    forall m.
+    (Monad m) =>
+    Maybe [a] ->
+    ProcessT m [a] [a]
   sortByBatchWith = \case
     Nothing -> MachineT $ pure $ Await onNext Refl onStop
      where
@@ -447,69 +458,146 @@ sortByBatch timestamp = sortByBatchWith Nothing
         | otherwise = MachineT $ pure $ Yield sortedBeforeCutoff $ sortByBatchWith $ Just sortedAfterCutoff
        where
         -- NOTE: use of partial @maximum@ is guarded by the check @null old@.
-        cutoff = getMax (foldMap (Max . timestamp) sortedOld)
+        cutoff = getMax (foldMap (Max . key) sortedOld)
         sortedNew = sortByTime new
         sorted = joinByTime sortedOld sortedNew
-        (sortedBeforeCutoff, sortedAfterCutoff) = L.partition ((<= cutoff) . timestamp) sorted
+        (sortedBeforeCutoff, sortedAfterCutoff) = L.partition ((<= cutoff) . key) sorted
       onStop :: ProcessT m [a] [a]
       onStop = MachineT $ pure $ Yield sortedOld $ stopped
 
-  -- compByTime :: a -> a -> Ordering
-  compByTime = compare `on` timestamp
+  compByTime :: a -> a -> Ordering
+  compByTime = compare `on` key
 
-  -- sortByTime :: [a] -> [a]
+  sortByTime :: [a] -> [a]
   sortByTime = L.sortBy compByTime
 
-  -- joinByTime :: [a] -> [a] -> [a]
+  joinByTime {- Sorted -} :: [a {- Sorted -}] -> [a {- Sorted -}] -> [a]
   joinByTime [] ys = ys
   joinByTime xs [] = xs
-  joinByTime (x : xs) (y : ys) = case compByTime x y of
-    LT -> x : joinByTime xs (y : ys)
-    _ -> y : joinByTime (x : xs) ys
+  joinByTime (x : xs) (y : ys)
+    | compByTime x y == LT = x : joinByTime xs (y : ys)
+    | otherwise = y : joinByTime (x : xs) ys
 
 {- |
-Variant of `sortByBatch` that operates on streams of items and ticks.
+Sort items between successive ticks.
+
+If the maximum key in batch @i@ is guaranteed to be smaller than the minimum
+key in batch @i + 2@, this process produces a totally ordered stream of items.
+
+==== __Examples__
+
+>>> run $ sortByTick @Int id <~ source [Item 1,Item 4,Tick,Item 7,Item 2,Item 3,Item 5,Tick,Item 6,Item 8]
+[Tick,Item 1,Item 2,Item 3,Item 4,Tick,Item 5,Item 6,Item 7,Item 8]
+
+>>> run $ sortByTick @Int id <~ source [Item 1,Item 7,Tick,Item 4,Item 2,Item 3,Item 5,Tick,Item 6,Item 8]
+[Tick,Item 1,Item 2,Item 3,Item 4,Item 5,Item 7,Tick,Item 6,Item 8]
 -}
-sortByBatchTick :: (a -> Timestamp) -> Process (Tick a) (Tick a)
-sortByBatchTick timestamp =
-  mapping (fmap (: [])) ~> batchByTick ~> sortByBatch timestamp ~> batchListToTick
+sortByTick ::
+  forall a k.
+  (Bounded k, Ord k) =>
+  (a -> k) ->
+  Process (Tick a) (Tick a)
+sortByTick key =
+  mapping (fmap D.singleton)
+    ~> batchByTick
+    ~> mapping (fmap D.toList)
+    ~> liftTick (sortByBatch key ~> asParts)
 
 -------------------------------------------------------------------------------
 -- Filtering semaphores
 -------------------------------------------------------------------------------
 
-{- | A simple delimiting t'Moore' machine,
-which is opened by one constant marker and closed by the other one.
--}
-between :: Text -> Text -> Moore Text Bool
-between x y = open
- where
-  open = Moore False open' where open' x' = if x == x' then close else open
-  close = Moore True close' where close' y' = if y == y' then end else close
-  end = Moore False (const end)
+{- |
+A delimiting t`Moore` machine based on constant open/close markers.
 
--- | Delimit the event process.
-delimit :: (Monad m) => Moore Text Bool -> ProcessT m Event Event
+The machine @`between` o c@ consumes consumes a stream of items, and produces
+a stream of `Bool` that is `False` up to and including the first occurrence of
+@o@, then is `True` up to and including the first occurrence of @c@, and then
+is `False` forever.
+
+==== __Examples__
+
+>>> run $ auto (betweenEach (2, 4)) <~ source [1, 2, 3, 4, 5]
+[False,False,True,True,False,False]
+
+>>> run $ auto (betweenEach (2, 4)) <~ source [2, 3, 4, 2, 3, 4]
+[False,True,True,False,True,True,False]
+-}
+betweenEach :: (Eq a) => (a, a) -> Moore a Bool
+betweenEach (open, close) = beforeOpen
+ where
+  beforeOpen = Moore False $ \a ->
+    if a == open then betweenOpenAndClose else beforeOpen
+  betweenOpenAndClose = Moore True $ \a ->
+    if a == close then beforeOpen else betweenOpenAndClose
+
+{- |
+A delimiting t`Moore` machine based on constant open/close markers.
+
+The machine @`between` o c@ consumes consumes a stream of items, and produces
+a stream of `Bool` that is `False` up to and including the first occurrence of
+@o@, then is `True` up to and including the first occurrence of @c@, and then
+is `False` forever.
+
+==== __Examples__
+
+>>> run $ auto (betweenFirst (2, 4)) <~ source [1, 2, 3, 4, 5]
+[False,False,True,True,False,False]
+
+>>> run $ auto (betweenFirst (2, 4)) <~ source [2, 3, 4, 2, 3, 4]
+[False,True,True,False,False,False,False]
+-}
+betweenFirst :: (Eq a) => (a, a) -> Moore a Bool
+betweenFirst (open, close) = beforeFirstOpen
+ where
+  beforeFirstOpen = Moore False $ \a ->
+    if a == open then betweenFirstOpenAndClose else beforeFirstOpen
+  betweenFirstOpenAndClose = Moore True $ \a ->
+    if a == close then afterFirstClose else betweenFirstOpenAndClose
+  afterFirstClose = Moore False (const afterFirstClose)
+
+{- |
+Filter the items in a stream based on a t`Moore` machine.
+
+==== __Examples__
+
+>>> run $ delimit (betweenEach (2, 4)) <~ source [1, 2, 3, 4, 5]
+[2,3,4]
+
+>>> run $ delimit (betweenEach (2, 4)) <~ source [2, 3, 4, 2, 3, 4]
+[2,3,4,2,3,4]
+
+>>> run $ delimit (betweenFirst (2, 4)) <~ source [1, 2, 3, 4, 5]
+[2,3,4]
+
+>>> run $ delimit (betweenFirst (2, 4)) <~ source [2, 3, 4, 2, 3, 4]
+[2,3,4]
+-}
+delimit ::
+  forall m a.
+  (Monad m) =>
+  Moore a Bool ->
+  ProcessT m a a
 delimit = construct . go
  where
-  go :: (Monad m) => Moore Text Bool -> PlanT (Is Event) Event m ()
-  go mm@(Moore s next) = do
-    e <- await
-    case evSpec e of
-      -- on marker step the moore machine.
-      E.UserMarker m -> do
-        let mm'@(Moore s' _) = next m
-        -- if current or next state is open (== True), emit the marker.
-        when (s || s') $ yield e
-        go mm'
-
-      -- for other events, emit if the state is open.
-      _ -> do
-        when s $ yield e
-        go mm
+  go ::
+    Moore a Bool ->
+    PlanT (Is a) a m Void
+  go _st@(Moore wasOpen onNext) =
+    await >>= \a -> do
+      -- Feed the item to the delimiting Moore machine.
+      let st'@(Moore willBeOpen _) = onNext a
+      -- If the state has changed, i.e., @wasOpen /= willBeOpen@, then
+      -- the current item is a marker. All markers should be yielded.
+      let isMarker = wasOpen /= willBeOpen
+      when (wasOpen || isMarker) $ yield a
+      go st'
 
 -------------------------------------------------------------------------------
 -- Validation
+--
+-- TODO: These machines, or at least the error messages that they print, are
+--       specific to eventlog processing. Hence, they should be moved.
 -------------------------------------------------------------------------------
 
 {- |
@@ -551,9 +639,9 @@ that directs the user to check that the @--eventlog-flush-interval@ and the
 @--batch-interval@ flags are set correctly.
 -}
 validateOrder ::
-  (MonadIO m, Show a) =>
+  (MonadIO m, Ord k, Show a) =>
   Verbosity ->
-  (a -> Timestamp) ->
+  (a -> k) ->
   ProcessT m a x
 validateOrder verbosity timestamp
   | verbosityError >= verbosity = construct $ start Nothing